backup-service/requirements.txt
cryptography==41.0.7
google-auth==2.23.4
google-auth-oauthlib==1.1.0
google-auth-httplib2==0.1.1
google-api-python-client==2.108.0
python-crontab==3.0.0
python-dateutil==2.8.2
